Allow use of elasticsearch 7.14.0
We've been using 7.14.0 with an earlier version of chewy for some time.
In order to upgrade to Ruby 3 we need to upgrade chewy and allow use of
our current version of elasticsearch. The only spec failures were in
Chewy.client.transport.logger and Chewy.client.transport.tracer which
are broken for us currently. Hopefully chewy will add support for this
in the future.